+++ blackdoc/tests/data/rst.py
@@ -69,6 +69,23 @@ a code block with a different language:
 .. code:: sh
 
     find . -name "*.py"
+
+a code block with testcode:
+
+.. testsetup::
+
+    file = open(
+        "very_long_filepath",
+        mode="a"
+    )
+
+.. testcode::
+
+    file
+
+.. testcleanup::
+
+    file.close()
 """
 lines = content.splitlines()
 line_ranges = (
@@ -122,6 +139,22 @@ line_ranges = (
     (66, 67),
     (67, 70),
     (70, 71),
+    (71, 72),
+    (72, 73),
+    (73, 74),
+    (74, 75),
+    (75, 76),
+    (76, 77),
+    (77, 78),
+    (78, 79),
+    (79, 80),
+    (80, 81),
+    (81, 82),
+    (82, 83),
+    (83, 84),
+    (84, 85),
+    (85, 86),
+    (86, 87),
 )
 line_labels = (
     "none",
@@ -140,6 +173,22 @@ line_labels = (
     "none",
     "none",
     "rst",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
+    "none",
     "none",
     "none",
     "none",